Fixed error in input validator and special characters
Now the program admits one only value and if this is not a letter you need to put another one.
te specials characters like accents are adding whit the vocals like a package. On this way if you insert 'A' the system recognize A and 'Á', and if you insert 'É' the system recognizes 'É' and 'E'.